cluster: add cluster level metrics via a new probe
This patch introduces a new probe that which is embedded in the
controller object. Only one instance is created per broker and it
matches the lifetime of the controller object. These metrics are
reported only from the controller leader (one node per cluster).

The following cluster level metrics were added:
* redpanda_cluster_brokers:
    * Description: number of configured brokers in the cluster
    * Label: shard
    * Aggreation: shard
* redpanda_cluster_partitions:
    * Description: number of partitions managed by the cluster (note
      that replicas are not counted as partitions and that the
      partitions of the controller topic are included)
    * Label: shard
    * Aggregation: shard
* redpanda_cluster_unavailable_partitions:
    * Description: number of unavailable partitions in the cluster (i.e.
      partitions that lack quorum among replicants)
    * Label: shard
    * Aggregation: shard