pandaproxy: Add latency metric to public endpoint
This patch adds request latency metrics for pandaproxy and the schema
registry.

* repdanda_pandaproxy_request_latency_seconds
* redpanda_schema_registry_request_latency_seconds
    * Description: latency of requests processed by the respective
      server which includes the time spent waiting for resources to
      become available
    * Labels: shard, operation
    * Aggregation: shard, operation

This patch changes the server intialisation to thread through the name
of the server to the metric registration scope. Prior to this change
pandaproxy and schema_registry metrics where both reported under the
`redpanda_pandaproxy` prefix, while now they're reported separately.