storage: use segment index for timequeries
Previously, the segment index was not used when performing a time-query.
This meant that the headers for all record batches in the segment had
to be read until the 'needle > batch_max_timestamp' condition was met.

If the batches within the segment have monotonical timestamps, the index
will be used. Otherwise, the old behaviour is kept. The consequence of
this is that timequery semantics are very clear: "a timequery for
timestamp 'needle' will return the first offset where 'timestamp >=
needle'".