storage: offset time index for out of order stamps
This commit introduces ofsetting (by 2^31) to the relative time index
stored in `index_state`. This ofsetting allows us to represent negative
deltas which can occur when dealing with client provided timestamps.
The rough range for these timestamps is from -596h to +596h.

Note that we need to apply the ofsetting when reading indices that are
already present on disk (both in the ADL and Serde cases).

The mapping looks as follows. Negative values are not drawn as they
could not exist with the previous indexing scheme.
Before ofsetting: [0, ..., 2 ^ 31 - 1, ..., 2 ^ 32 - 1]
                   |             |              |
                   |             |______________|
                   |             |
After ofsetting:  [2^31, ..., 2 ^ 32 - 1]