cluster: RPC for fetching cloud log size on shard
This commit introduces a new RPC, `get_partitions_cloud_storage_size`.
The request provides a list of partitions, while the reply provides the
total cloud log size of the partitions that were on the shard at the
time of processing the request. Missing partitions are also included in
the reply.